[feat] 도메인 예외 생성
