--- src/main/java/com/hhconcert/server/global/common/exception/ConcertException.java
@@ -1,8 +1,0 @@
-package com.hhconcert.server.global.common.exception;
-
-public class ConcertException extends RuntimeException {
-
-    public ConcertException(String message) {
-        super(message);
-    }
-}
