--- src/main/java/com/hhconcert/server/global/common/exception/PaymentException.java
@@ -1,7 +1,0 @@
-package com.hhconcert.server.global.common.exception;
-
-public class PaymentException extends RuntimeException {
-    public PaymentException(String message) {
-        super(message);
-    }
-}
