--- src/main/java/com/hhconcert/server/global/common/exception/PointException.java
@@ -1,7 +1,0 @@
-package com.hhconcert.server.global.common.exception;
-
-public class PointException extends RuntimeException {
-    public PointException(String message) {
-        super(message);
-    }
-}
