--- src/main/java/com/hhconcert/server/global/common/exception/ReservationException.java
@@ -1,7 +1,0 @@
-package com.hhconcert.server.global.common.exception;
-
-public class ReservationException extends RuntimeException {
-    public ReservationException(String message) {
-        super(message);
-    }
-}
