--- src/main/java/com/hhconcert/server/global/common/exception/TokenException.java
@@ -1,7 +1,0 @@
-package com.hhconcert.server.global.common.exception;
-
-public class TokenException extends RuntimeException {
-    public TokenException(String message) {
-        super(message);
-    }
-}
